Initial CartesianSeries generic param  refactor.
